# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Heyya school chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{roomId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in a room
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: number
        - in: query
          name: before
          schema:
            type: string
            description: Cursor – ISO timestamp of oldest message already loaded
        - in: query
          name: limit
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a room
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /rooms/{roomId}/messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      summary: Add an emoji reaction to a message
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: number
        - in: path
          name: messageId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Attendance ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, teacher, student, parent]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, on_leave, withdrawn]
        childId:
          type: ["number", "null"]
          description: Parent's linked student ID
        createdAt:
          type: string
          format: date-time
...
    Message:
      type: object
      required: [id, roomId, type, content, createdAt]
      properties:
        id:
          type: number
        roomId:
          type: number
        senderId:
          type: ["number", "null"]
        sender:
          $ref: "#/components/schemas/User"
        type:
          type: string
          enum: [text, system, attendance_card, leave_card, ai_message, file, confirmation_card, hook_card]
        content:
          type: string
        metadata:
          type: ["object", "null"]
          description: Extra data for card-type messages (attendance, leave, etc.)
          additionalProperties: true
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum: [text, system, attendance_card, leave_card, ai_message, file, confirmation_card, hook_card]
          default: text
        metadata:
          type: object
          additionalProperties: true
    Reaction:
      type: object
      required: [emoji, userIds]
      properties:
        emoji:
          type: string
        userIds:
          type: array
          items:
            type: number
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string

    # ── Attendance ────────────────────────────────────────────────────────────
...